Fixing doc typos and model refactoring
