+++ client/src/rubin/nublado/client/models/__init__.py
@@ -5,13 +5,16 @@ from ._image import (
     NubladoImageClass,
     NubladoImageSize,
 )
+from ._jupyter import JupyterOutput, SpawnProgressMessage
 from ._user import User
 
 __all__ = [
+    "JupyterOutput",
     "NotebookExecutionResult",
     "NubladoImage",
     "NubladoImageClass",
     "NubladoImageByClass",
     "NubladoImageSize",
+    "SpawnProgressMessage",
     "User",
 ]
